Delete and move files
