--- shared/src/androidMain/kotlin/com/corrado4eyes/cucumberplayground/Platform.kt
@@ -1,7 +1,0 @@
-package com.corrado4eyes.cucumberplayground
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
