--- shared/src/commonMain/kotlin/com/corrado4eyes/cucumberplayground/Greeting.kt
@@ -1,9 +1,0 @@
-package com.corrado4eyes.cucumberplayground
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, this is CucumberPlayground!"
-    }
-}
