--- shared/src/commonMain/kotlin/com/corrado4eyes/cucumberplayground/Platform.kt
@@ -1,7 +1,0 @@
-package com.corrado4eyes.cucumberplayground
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
