--- shared/src/iosMain/kotlin/com/corrado4eyes/cucumberplayground/Platform.kt
@@ -1,9 +1,0 @@
-package com.corrado4eyes.cucumberplayground
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = "this is iOS 🍎"
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
